requirements.txt
streamlit
python-dotenv==1.0.1
Pillow==9.5.0
google-generativeai==0.0.2